fix(ds): make the functions common
